requirements.txt
torch
pandas
Pillow
matplotlib
torchvision
tqdm